Initial work on decoupling message decrypting/deserialization from ownership of the websocket.
