+++ auxin_cli/src/lib.rs
